requirements.txt
Django==2.1.5
django-debug-toolbar==1.10.1
Pillow==8.0.1
python-decouple==3.1
pytz==2018.5
sqlparse==0.2.4
